Changed formatting
improved playback on songs, miniplayer is fully functional, Slider is no longer an issue, playlist is fully functional, Navbar's style changed to match figma and the mini player. Favourites is still not fully functioning, it can store songs but can't play from the page. The log in and authentication has been removed so it would be easier to implement davonte's. The profile page is static for now until log in is fully implemented with db.